pyproject.toml
[tool.poetry]
name = "cards-api"
version = "0.1.0"
description = ""
authors = ["Alexander Payne <81712870+DrPayne25@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0.3"
djangorestframework = "^3.13.1"
psycopg2-binary = "^2.9.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
